myproject/requirements.txt

Django==4.2.16
django-cors-headers==4.4.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.7
et-xmlfile==1.1.0
fastapi==0.110.2
filelock==3.13.4
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.25.0
gunicorn==20.1.0
h11==0.14.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.3
jsonpointer==2.1
kombu==5.4.2
MarkupSafe==2.1.5
munkres==1.1.4
oauthlib==3.2.2
openpyxl==3.0.10
patsy==0.5.3
pep8==1.7.1
pipenv==2024.0.1
ply==3.11
psycopg2-binary==2.9.9
pyarrow==11.0.0
pyasn1-modules==0.2.8
PyDispatcher==2.0.5
PyJWT==1.7.1
pyls-spyder==0.4.0
PyQt5-sip==12.11.0
pyTelegramBotAPI==4.16.1
python-lsp-jsonrpc==1.0.0
python3-openid==3.2.0
queuelib==1.5.0
redis==4.6.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
SQLAlchemy==2.0.31
sqlparse==0.5.1
starlette==0.37.2
telebot==0.0.5
typing_extensions==4.11.0
uritemplate==4.1.1
uvicorn==0.29.0
vine==5.1.0
virtualenv==20.26.0
webencodings==0.5.1
Werkzeug==3.0.2
WTForms==3.1.2
